% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{blca_clinical}
\alias{blca_clinical}
\title{Bladder Cancer Clinical Data}
\format{
\describe{
\item{ids}{sample ID}
\item{status}{sample survival status (1 = dead,0 = alive)}
\item{time}{survival time (days)}
\item{Gender}{the sample gender}
\item{...}{other clinical informations}
}
}
\usage{
blca_clinical
}
\description{
the clinical information (including: time, status...) of bladder cancer
}
