% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{combn_signature}
\alias{combn_signature}
\title{get all combination of specific number}
\usage{
combn_signature(genes, n)
}
\arguments{
\item{genes}{the candidate set of genes}

\item{n}{the number of genes to select}
}
\value{
return a list of combination, each row is a combination
}
\description{
get all combination of specific number
}
\examples{
combn_signature(letters[1:4], n = 2)
combn_signature(letters[1:4], n = 2:3)

}
