% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{get_beta}
\alias{get_beta}
\title{beta coefficient of univariate cox analysis}
\usage{
get_beta(run_cox_obj)
}
\arguments{
\item{run_cox_obj}{the \code{run_cox} object get from the \link{run_cox}()}
}
\value{
return a \code{data.frame}, the first column is \code{Variable}, the second column is \code{beta}.
}
\description{
get the beta coefficient of univariate cox analysis
}
\examples{
library(survival)
data(lung)
cox_obj <- run_cox(data = lung, time = "time", event = "status",
  variate = c("age", "sex", "ph.ecog"), multicox = FALSE, global_method = "wald")
get_beta(cox_obj)

}
