% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{label_sample}
\alias{label_sample}
\title{High and Low}
\usage{
label_sample(score, value = NULL)
}
\arguments{
\item{score}{the vector of risk score.}

\item{value}{the specific value to distinguish different samples. Default is \code{NULL},
we used the median of risk score to divide samples.}
}
\description{
Differentiate samples based on risk score
}
\details{
If the score is higher than the specific value, the samples are labeled
as \code{High_risk}, otherwise as \code{Low_risk}
}
\examples{
label_sample(1:5)
label_sample(1:5, value = 3)
}
