% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untilis.R
\name{logrank_p}
\alias{logrank_p}
\title{calculate logrank test p value}
\usage{
logrank_p(data, time = "time", event = "status", variate)
}
\arguments{
\item{data}{a \code{data.frame} containing variables, time and status.}

\item{time}{column name specifying time, default is 'time'.}

\item{event}{column name specifying event status, default is 'status'.}

\item{variate}{column names of specific variables in the survival model.}
}
\value{
return the logrank test p value
}
\description{
calculate logrank test p value
}
\examples{
library(survival)
data(ovarian)
logrank_p(data = ovarian, time = "futime", event = "fustat", variate = "rx")
logrank_p(data = ovarian, time = "futime", event = "fustat", variate = c("rx", "resid.ds"))
}
