% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{meta_dat}
\alias{meta_dat}
\title{A special case of data}
\format{
\describe{
\item{ids}{sample ID}
\item{time}{survival time (days)}
\item{status}{sample survival status (1 = dead,0 = alive)}
\item{...}{gene expression value in different samples}
}
}
\usage{
meta_dat
}
\description{
a special case of data run \link{risk_score}() function
}
