% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{risk_score}
\alias{risk_score}
\title{calculate the expression risk score}
\usage{
risk_score(exp_data, genes, beta)
}
\arguments{
\item{exp_data}{a \code{matrix} of expression profile.}

\item{genes}{a character vector contain the genes in the signature.}

\item{beta}{a  contained the beta value
correspond to the genes.}
}
\value{
return a \code{tibble}
}
\description{
batch calculate the expression risk score
}
\details{
this function calculate the risk score based on expression profile.
\verb{risk score = gene_1 * coef_1 + ... + gene_n * coef_n}
}
\examples{
\dontrun{
    data(meta_dat)
    data(blca_exp)
    genes <- colnames(meta_dat)[4:10]
    cox_res <- run_cox(data = meta_dat, time = "time", event = "status",
                       variate = genes, multicox = FALSE, global_method = "wald")
    beta <- get_beta(cox_res)
    scores <- risk_score(exp_data = blca_exp, genes, beta)
}

}
