% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{run_cox_parallel}
\alias{run_cox_parallel}
\title{Run Cox Analysis in Batch Mode}
\usage{
run_cox_parallel(
  data,
  time = "time",
  event = "status",
  variate,
  multicox = FALSE,
  global_method = c("likelihood", "wald", "logrank")
)
}
\arguments{
\item{data}{a \code{data.frame} containing variables, time and event.}

\item{time}{column name specifying time, default is 'time'.}

\item{event}{column name specifying event status, default is 'status'.}

\item{variate}{column names specifying variables.}

\item{multicox}{default \code{FALSE}. If \code{TRUE}, do the multiple variate cox analysis.}

\item{global_method}{method used to obtain global p value for cox model,
should be one of "likelihood", "wald", "logrank".
The likelihood-ratio test, Wald test, and score logrank statistics.
These three methods are asymptotically equivalent. Default is "wald".}
}
\value{
return a \code{run_cox} object including the summary information of cox analysis
}
\description{
Run Cox Analysis in Batch Mode
}
\examples{
library(survival)
data(lung)
run_cox_parallel(data = lung, time = "time", event = "status",
  variate = c("age", "sex", "ph.ecog"), multicox = FALSE, global_method = "wald")
run_cox_parallel(data = lung, time = "time", event = "status",
  variate = c("age", "sex", "ph.ecog"), multicox = TRUE, global_method = "wald")

}
