% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivals.R
\name{train_multicox}
\alias{train_multicox}
\title{Do the multivariate cox analysis in training data sets.}
\usage{
train_multicox(
  obj,
  type = c("continuous", "discrete"),
  covariate = NULL,
  cut_p = 0.05
)
}
\arguments{
\item{obj}{the \code{training_signature} object get from \link{train_signature}() .}

\item{type}{Use which variate to do the multivariate cox analysis, if \code{continuous}, use the \code{risk_score};
if \code{discrete}, use the \code{labels}.}

\item{cut_p}{the cutoff p value of multivariate cox analysis. Default 0.05.}
}
\value{
return a \code{training_signature} object with \code{multicox_pval} column.
}
\description{
Do the multivariate cox analysis in training data sets.
}
\examples{
\dotrun{
  multi_cox_obj <- train_multicox(obj = uni_cox_obj, type = "discrete", covariate = c("Age", "Gender"), cut_p = 1)
  multi_cox_obj <- train_multicox(obj = obj, type = "discrete", covariate = c("Age", "Gender"), cut_p = 1)
  ## covariate = NULL
  multi_cox_obj <- train_multicox(obj = uni_cox_obj, type = "discrete", cut_p = 1)
}
}
