% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivals.R
\name{train_signature}
\alias{train_signature}
\title{Train the signature model in a data sets}
\usage{
train_signature(
  surv_data,
  id = "ids",
  time = "time",
  event = "status",
  exp_data,
  genes,
  num,
  beta = NULL,
  labels_value = NULL,
  cut_p = NULL,
  keep_data = TRUE
)
}
\arguments{
\item{surv_data}{a \code{data.frame} containing variables, ids, time and event.}

\item{id}{column name specifying samples ID , default is 'ids'.}

\item{time}{column name specifying time, default is 'time'.}

\item{event}{column name specifying event status, default is 'status'.}

\item{exp_data}{a \code{matrix} of expression profile. row names are genes, column names are sample ID.}

\item{genes}{a character vector of genes to construct risk model. They must be in \code{exp_data} row names.}

\item{num}{a numeric vector of the number of genes in the risk model.}

\item{beta}{if \code{NULL}, the function compute the beta value for each gene in univariate cox analysis;
or a data like the return object of \link{get_beta}()}

\item{labels_value}{the same like \code{value} in \link{label_sample}()}

\item{cut_p}{a p value to filter the result of \code{train_signature}, we only keep the risk model with p value
samller than the value}

\item{keep_data}{whether to keep the data used to compute logrank test in this function,
It may be very important for follow-up analysis.}
}
\description{
this function will filter the signature are not significant in the logrank test
}
\examples{
\dotrun{
data(blca_clinical)
data(blca_exp)
obj <- train_signature(surv_data = blca_clinical, id = "ids", time = "time", event = "status",
                       exp_data = blca_exp, genes = c("IL6", "TGFB3", "VHL", "CXCR4"), num = 2:4,
                       beta = NULL, labels_value = NULL, cut_p = 1, keep_data = T)
}

}
