% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivals.R
\name{train_unicox}
\alias{train_unicox}
\title{Do the univariate cox analysis in training data sets.}
\usage{
train_unicox(obj, type = c("continuous", "discrete"), cut_p = 0.05)
}
\arguments{
\item{obj}{the \code{training_signature} object get from \link{train_signature}() .}

\item{type}{Use which variate to do the univariate cox analysis, if \code{continuous}, use the \code{risk_score};
if \code{discrete}, use the \code{labels}.}

\item{cut_p}{the cutoff p value of univariate cox analysis. Default 0.05.}
}
\value{
return a \code{training_signature} object with \code{unicox_pval} column.
}
\description{
Do the univariate cox analysis in training data sets.
}
