% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivals.R
\name{validate_signature}
\alias{validate_signature}
\title{Validate the signature model in a data sets}
\usage{
validate_signature(
  train_sig = NULL,
  surv_data,
  id = "ids",
  time = "time",
  event = "status",
  exp_data,
  cut_p = NULL
)
}
\value{

}
\description{
Validate the signature model in a data sets
}
\examples{

}
\author{

}
