% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untilis.R
\name{logrank_p}
\alias{logrank_p}
\title{calculate logrank test p value}
\usage{
logrank_p(data, time = "time", event = "status", variate, verbose = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} containing variables, time and status.}

\item{time}{column name specifying time, default is 'time'.}

\item{event}{column name specifying event status, default is 'status'.}

\item{variate}{column names of specific variables in the survival model.}

\item{verbose}{output other useful information, default TRUE.}
}
\value{
return a tibble object. The first column is the variate name,
the second column is the logrank test p value, the third column is the Survival Object.
}
\description{
calculate logrank test p value
}
\note{
If there is some wrong with the logrank test, the function will set logrank p value to 1 automatically.
So set \code{verbose} argument to TRUE is necessary to check these events.
}
\examples{
library(survival)
data(ovarian)
logrank_p(data = ovarian, time = "futime", event = "fustat", variate = "rx")
logrank_p(data = ovarian, time = "futime", event = "fustat", variate = c("rx", "resid.ds"))
}
